services/tekup-gmail-services/apps/gmail-automation/requirements.txt
# Core dependencies
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.5.0
google-auth>=2.0.0
requests>=2.28.0
python-dotenv>=0.19.0
pydantic>=1.10.0
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
aiofiles>=0.8.0

# Image processing
pillow>=9.0.0
pytesseract>=0.3.8
opencv-python>=4.6.0
numpy>=1.21.0

# Data processing
pandas>=1.4.0
openpyxl>=3.0.0

# Scheduling
schedule>=1.2.0
python-dateutil>=2.8.0
pytz>=2022.1

# Email
email-validator>=1.2.0

# Templating
jinja2>=3.1.0
markdown>=3.4.0

# Configuration
pyyaml>=6.0
toml>=0.10.2

# CLI
click>=8.1.0
rich>=12.5.0

# Logging
loguru>=0.6.0

# HTTP
httpx>=0.23.0
aiohttp>=3.8.0
websockets>=10.0

# Database
redis>=4.3.0
celery>=5.2.0
sqlalchemy>=1.4.0
alembic>=1.8.0
psycopg2-binary>=2.9.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.19.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
isort>=5.10.0
flake8>=4.0.0
mypy>=0.950
pre-commit>=2.17.0

# Retry logic
tenacity>=8.0.0